+++ src/index.html
@@ -50,10 +50,6 @@
             </div>
           </div> <!-- fin del filtros-->
           <div id="todos"></div>
-          <!-- <div id="soloG"></div>
-          <div id="soloH"></div>
-          <div id="soloR"></div>
-          <div id="soloS"></div> -->
         </div> <!-- fin de menu Casas-->
 
         <!-- Menu Rol -->
